Move code for shared use of the fragment dialog
