+++ avogadro/qtgui/insertfragmentdialog.h
