+++ avogadro/qtgui/insertfragmentdialog.ui
