+++ avogadro/qtgui/sortfiltertreeproxymodel.h
